+++ package.json
@@ -16,7 +16,6 @@
     "@taquito/tzip16": "^14.0.0",
     "@taquito/utils": "^14.0.0",
     "axios": "^1.0.0",
-    "concurrently": "^8.2.2",
     "dotenv": "^16.0.3",
     "express": "^4.18.1",
     "node-fetch": "^3.2.4",
